Moving acm fixtures out of yaml editor
