+++ packages/nxtcm-rosa-hcp-wizard/src/ROSAHCPWizard.stories.tsx
@@ -16,12 +16,13 @@ import {
 } from './ROSAHCPWizard.stories.helpers';
 import fixtures from './ROSAHCPWizard.fixtures';
 import { MachineTypesDropdownType, OIDCConfig, Region, Role, ROSAHCPWizardData } from './types';
+import { STEP_IDS } from './constants';
 import { defaultRosaHcpWizardStrings } from './stringsProvider/rosaHcpWizardStrings.defaults';
-import { createAcmCapaGenerator } from './Steps/YamlEditor/generators/acmCapaGenerator';
-import rosaControlPlaneSchema from './Steps/YamlEditor/schemas/rosaControlPlaneSchema.json';
-import managedClusterSchema from './Steps/YamlEditor/schemas/managedClusterSchema.json';
-import capiClusterSchema from './Steps/YamlEditor/schemas/capiClusterSchema.json';
-import rosaClusterSchema from './Steps/YamlEditor/schemas/rosaClusterSchema.json';
+import { createAcmCapaGenerator } from './test/acmGeneratorFixtures/acmCapaGenerator';
+import rosaControlPlaneSchema from './test/acmGeneratorFixtures/schemas/rosaControlPlaneSchema.json';
+import managedClusterSchema from './test/acmGeneratorFixtures/schemas/managedClusterSchema.json';
+import capiClusterSchema from './test/acmGeneratorFixtures/schemas/capiClusterSchema.json';
+import rosaClusterSchema from './test/acmGeneratorFixtures/schemas/rosaClusterSchema.json';
 
 const storyAcmGenerator = createAcmCapaGenerator([
   { kind: 'ROSAControlPlane', schema: rosaControlPlaneSchema, primary: true },
@@ -399,6 +400,28 @@ export const RolesAlertMissingUserRole: Story = {
         fetch: async () => {},
       },
     }),
+  },
+};
+
+export const AcmConfiguration: Story = {
+  render: (args) => <DefaultStoryWrapper {...args} />,
+  args: {
+    title: 'Create ROSA Cluster — ACM',
+    resourceGenerator: storyAcmGenerator,
+    product: 'acm',
+    wizardData: createMockRosaHcpWizardDataWithFetchLogging(),
+    config: {
+      hiddenSteps: [STEP_IDS.CLUSTER_WIDE_PROXY, STEP_IDS.CLUSTER_UPDATES],
+    },
+    onSubmit: async (data: unknown) => {
+      console.log('Wizard submitted with data:', data);
+      await new Promise((resolve) => setTimeout(resolve, 2000));
+      alert('Cluster creation initiated successfully!');
+    },
+    onCancel: () => {
+      console.log('Wizard cancelled');
+      alert('Wizard cancelled');
+    },
   },
 };
