+++ packages/nxtcm-rosa-hcp-wizard/src/ROSAHCPWizardBody.tsx
@@ -73,7 +73,6 @@ export const ROSAHCPWizardBody = (props: RosaHCPWizardProps) => {
   const clusterWideProxySelected = useWatch({ name: 'configure_proxy' });
 
   const isClusterWideProxyHidden = useIsStepHidden(STEP_IDS.CLUSTER_WIDE_PROXY);
-  const isEncryptionHidden = useIsStepHidden(STEP_IDS.ENCRYPTION);
   const isClusterUpdatesHidden = useIsStepHidden(STEP_IDS.CLUSTER_UPDATES);
 
   const rosaStrings = useRosaHcpWizardStrings();
@@ -156,15 +155,13 @@ export const ROSAHCPWizardBody = (props: RosaHCPWizardProps) => {
 
         {(() => {
           const optionalSteps = [
-            !isEncryptionHidden && (
-              <WizardStep
-                name={sl.encryptionOptional}
-                id={STEP_IDS.ENCRYPTION}
-                key={STEP_IDS.ENCRYPTION}
-              >
-                <Encryption />
-              </WizardStep>
-            ),
+            <WizardStep
+              name={sl.encryptionOptional}
+              id={STEP_IDS.ENCRYPTION}
+              key={STEP_IDS.ENCRYPTION}
+            >
+              <Encryption />
+            </WizardStep>,
             !isClusterUpdatesHidden && (
               <WizardStep
                 name={sl.clusterUpdatesOptional}
