--- packages/nxtcm-rosa-hcp-wizard/src/Steps/YamlEditor/generators/index.ts
@@ -1,4 +1,0 @@
-export { createAcmCapaGenerator } from './acmCapaGenerator';
-export { createTemplateBasedGenerator } from './createTemplateBasedGenerator';
-export type { TemplateBasedGeneratorOptions } from './createTemplateBasedGenerator';
-export { createOcmClusterServiceGenerator } from './ocmClusterServiceGenerator';
