+++ packages/nxtcm-rosa-hcp-wizard/src/test/acmGeneratorFixtures/acmCapaGenerator.ts
@@ -1,9 +1,9 @@
 import Handlebars from 'handlebars';
 
-import { parseRosaControlPlaneYaml } from '../yamlUtils';
-import type { YamlResourceGenerator, ResourceSchema } from '../types';
+import { parseRosaControlPlaneYaml } from '../../Steps/YamlEditor/yamlUtils';
+import type { YamlResourceGenerator, ResourceSchema } from '../../Steps/YamlEditor/types';
 import { createTemplateBasedGenerator } from './createTemplateBasedGenerator';
-import rosaHcpTemplateRaw from '../templates/rosa-hcp-template.hbs?raw';
+import rosaHcpTemplateRaw from './templates/rosa-hcp-template.hbs?raw';
 
 const eqHelper: Handlebars.HelperDelegate = function (
   this: unknown,
