+++ packages/nxtcm-rosa-hcp-wizard/src/test/acmGeneratorFixtures/schemas/rosaClusterSchema.json
