+++ packages/nxtcm-rosa-hcp-wizard/src/test/acmGeneratorFixtures/schemas/rosaControlPlaneSchema.json
