+++ packages/nxtcm-rosa-hcp-wizard/src/test/acmGeneratorFixtures/templates/rosa-hcp-template.hbs
